Also move the PolarisTreeMap* classes into the transactional package
